scraper/review_extractor.py
"""
Review extraction utilities for Amazon scraper
"""

from bs4 import BeautifulSoup
import re
from typing import Dict, List, Optional

from utils.data_utils import clean_text, extract_rating


class ReviewExtractor:
    """Extract review information from Amazon pages"""
    
    @staticmethod
    def extract_from_review_page(element) -> Optional[Dict]:
        """Extract review from review element"""
        try:
            # This is a placeholder implementation
            # In a real implementation, this would parse HTML elements
            return {
                'review_id': 'sample_review_id',
                'rating': 5,
                'title': 'Great product',
                'text': 'Excellent quality and value'
            }
        except Exception:
            return None
