requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.1
pandas>=2.1.3
numpy>=1.24.3

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2
requests>=2.31.0

# Data analysis & visualization
plotly>=5.17.0
matplotlib>=3.8.2
seaborn>=0.13.0

# Text processing & NLP
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
scikit-learn>=1.3.2

# Database
sqlalchemy>=2.0.23

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Development & testing
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
